refactor: separate phoneInput component
