+++ client/src/components/Sign/index.js
@@ -1,2 +1,3 @@
 export { default as FormInput } from './FormInput';
 export { default as FormAddressInput } from './FormAddressInput';
+export { default as FormPhoneInput } from './FormPhoneInput';
